+++ test/features/account/domain/account_test.dart
@@ -1,7 +1,7 @@
 import 'package:anonaddy/features/account/domain/account.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_data/account_test_data.dart';
+import '../../../test_data/account_test_data.dart';
 
 void main() {
   group('Account model testing ', () {
