+++ test/features/account/presentation/account_tab_header_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../mocks.dart';
+import '../../../mocks.dart';
 
 void main() {
   // late MockAccountService accountService;
