+++ test/features/auth/data/mock_flutter_secure_storage.dart
@@ -5,7 +5,7 @@ import 'package:anonaddy/shared_components/constants/secure_storage_keys.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../test_data/account_test_data.dart';
+import '../../../test_data/account_test_data.dart';
 
 class MockFlutterSecureStorage extends Mock implements FlutterSecureStorage {
   @override
