+++ test/features/auth/presentation/loading_screen_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'auth_screen_mocks.dart';
+import '../../features/auth/presentation/auth_screen_mocks.dart';
 
 void main() {
   Widget loadingScreen() {
