+++ test/features/failed_deliveries/data/failed_delivery_service_test.dart
@@ -4,8 +4,8 @@ import 'package:dio/dio.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../features/failed_deliveries/data/dummy_failed_deliveries_data.dart';
 import '../../mocks.dart';
-import 'dummy_failed_deliveries_data.dart';
 
 void main() async {
   late MockDio mockDio;
